% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{checklegal}
\alias{checklegal}
\title{Check legal}
\usage{
checklegal(table, release)
}
\arguments{
\item{table}{"mean", "stack", "forced_mean", "detection"}

\item{release}{"dr2", "dr1"}
}
\description{
Checks if this combination of table and release is acceptable.
}
\keyword{internal}
