% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{convert_types}
\alias{convert_types}
\title{Convert types.}
\usage{
convert_types(dt, cols, types)
}
\arguments{
\item{dt}{data.table or data.frame}

\item{cols}{which cols convert}

\item{types}{"char", "long", "unsignedByte", "int", "short", "double", "float"}
}
\value{
a data frame with correct data types
}
\description{
Helper function.
}
\keyword{internal}
