% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{get_metadata}
\alias{get_metadata}
\title{Get preloaded metadata}
\usage{
get_metadata(table, release)
}
\arguments{
\item{table}{"mean", "stack", "detection"}

\item{release}{"dr2", "dr1"}
}
\value{
A data frame
}
\description{
Get preloaded metadata
}
\keyword{internal}
