% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{ps1_image_color}
\alias{ps1_image_color}
\title{Get color image at a sky position}
\usage{
ps1_image_color(
  ra,
  dec,
  size = 240,
  output_size = NULL,
  filters = "grizy",
  format = "jpg"
)
}
\arguments{
\item{ra}{ra position in degrees}

\item{dec}{dec position in degrees}

\item{size}{extracted image size in pixels (0.25 arcsec/pixel)}

\item{output_size}{output (display) image size in pixels (default = size).
output_size has no effect for fits format images.}

\item{filters}{string with filters to include}

\item{format}{data format (options are "jpg", "png")}
}
\value{
the image url
}
\description{
Get color image at a sky position
}
\examples{
\dontrun{
ps1_image_color(ra = 83.633210, dec = 22.014460, size = 1280, filters="grz")
}
}
