% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{ps1_image_gray}
\alias{ps1_image_gray}
\title{Get grayscale image at a sky position}
\usage{
ps1_image_gray(
  ra,
  dec,
  size = 240,
  output_size = NULL,
  filter = "g",
  format = "jpg"
)
}
\arguments{
\item{ra}{ra position in degrees}

\item{dec}{dec position in degrees}

\item{size}{extracted image size in pixels (0.25 arcsec/pixel)}

\item{output_size}{output (display) image size in pixels (default = size).
output_size has no effect for fits format images.}

\item{filter}{string with filter to extract (one of grizy)}

\item{format}{data format (options are "jpg", "png")}
}
\value{
the image
}
\description{
Get grayscale image at a sky position
}
\examples{
\dontrun{
ps1_image_gray(ra = 83.633210, dec = 22.014460, size = 1280, filter = "i")
}
}
