% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{ps1_image_list}
\alias{ps1_image_list}
\title{Get list of images}
\usage{
ps1_image_list(ra, dec, size = 240, filters = "grizy")
}
\arguments{
\item{ra}{ra position in degrees}

\item{dec}{dec position in degrees}

\item{size}{image size in pixels (0.25 arcsec/pixel)}

\item{filters}{string with filters to include}
}
\value{
table with the results
}
\description{
Query ps1filenames.py service to get a list of images.
}
\details{
src: https://ps1images.stsci.edu/ps1image.html
}
\examples{
\dontrun{
# Crab nebulae image
ps1_image_list(ra = 83.633210, dec = 22.014460, size = 1280, filters = "grz")
}

}
