% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{ps1_mast_query}
\alias{ps1_mast_query}
\title{Perform a MAST query.}
\usage{
ps1_mast_query(request)
}
\arguments{
\item{request}{The MAST request json object}
}
\value{
Returns response
}
\description{
Perform a MAST query.
}
\keyword{internal}
