% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{ps1_mast_resolve}
\alias{ps1_mast_resolve}
\title{Get the RA and Dec for an object using the MAST name resolver}
\usage{
ps1_mast_resolve(name)
}
\arguments{
\item{name}{Name of object}
}
\value{
list of ra, decl
}
\description{
Get the RA and Dec for an object using the MAST name resolver
}
\examples{
\dontrun{
ps1_mast_resolve('Acrux')
}
}
