% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{ps1_metadata}
\alias{ps1_metadata}
\title{Metadata from PS1}
\usage{
ps1_metadata(table = "mean", release = "dr2")
}
\arguments{
\item{table}{"mean", "stack", "forced_mean" or "detection"}

\item{release}{"dr1" or "dr2"(default)}
}
\value{
Returns data.frame with columns: name, type, description
}
\description{
Return metadata for the specified catalog and table
}
\examples{
\dontrun{
ps1_metadata()
}
}
