% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_catalog_search.R
\name{ps1_resolve}
\alias{ps1_resolve}
\title{Get the RA and Dec for objects from PanSTARRS catalog.}
\usage{
ps1_resolve(target_names, verbose = FALSE)
}
\arguments{
\item{target_names}{character vector of target names (see example)}

\item{verbose}{print info about request}
}
\value{
data.frame
}
\description{
Only works for "north" objects with decl > -30. For all objects see function `ps1_mast_resolve`.
}
\examples{
\dontrun{
ps1_resolve(c('Andromeda', "SN 2005D", 'Antennae', 'ANTENNAE'))
}
}
