% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_radec}
\alias{validate_radec}
\title{Check ra, dec params}
\usage{
validate_radec(ra, dec, .length = NULL)
}
\arguments{
\item{ra}{(degrees) Right Ascension}

\item{dec}{(degrees) Declination}

\item{.length}{length of coordinates. If `NULL` (default) then coordinates
can be any size.}
}
\value{
If the check is successful, the function returns nothing.
}
\description{
Check ra, dec params
}
\keyword{internal}
